+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.9.15] - UNRELEASED
 
 - Added help buttons on every dialog
+- Added --shortcut CLI option for pgtools.exe to keep it open when done running until ENTER is pressed
 - Refined warning and error messages to be more clear and reflect the [docs](https://github.com/hakasapl/PGPatcher/wiki/Error-Message-Guide)
 
 ## [0.9.14] - 2026-02-21